Tests: abstract the assertAttributeSame backfill out to a trait
... to allow use of this functionality in other tests.

At some point in the future this may be (should probably be) abstracted out altogether to a separate package or to the `TestUtils`, but I need to think about that a little more.
For now, setting it up as a `trait` solves the more immediate pressing issue.
