+++ Tests/AbstractSniffs/AbstractArrayDeclaration/AbstractArrayDeclarationSniffTest.php
@@ -10,10 +10,8 @@
 
 namespace PHPCSUtils\Tests\AbstractSniffs\AbstractArrayDeclaration;
 
-use Exception;
+use PHPCSUtils\Tests\AssertAttributeSame;
 use PHPCSUtils\TestUtils\UtilityMethodTestCase;
-use ReflectionException;
-use ReflectionObject;
 
 /**
  * Tests for the \PHPCSUtils\AbstractSniffs\AbstractArrayDeclarationSniff class.
@@ -26,6 +24,8 @@ use ReflectionObject;
  */
 class AbstractArrayDeclarationSniffTest extends UtilityMethodTestCase
 {
+    // Backfill for the assertAttributeSame() method on PHPUnit 9.x.
+    use AssertAttributeSame;
 
     /**
      * List of methods in the abstract which should be mocked.
@@ -639,73 +639,5 @@ class AbstractArrayDeclarationSniffTest extends UtilityMethodTestCase
             ->willReturn(true);
 
         $mockObj->process(self::$phpcsFile, $target);
-    }
-
-    /**
-     * PHPUnit cross-version helper method to test the value of the class properties.
-     *
-     * @param mixed  $expected      Expected property value.
-     * @param string $attributeName The name of the property to check.
-     * @param object $actualObject  The object on which to check the property value.
-     * @param string $message       Optional. Custom error message.
-     *
-     * @return void
-     */
-    public function assertAttributeValueSame($expected, $attributeName, $actualObject, $message = '')
-    {
-        // Will throw a warning on PHPUnit 8, but will still work.
-        if (\method_exists($this, 'assertAttributeSame')) {
-            parent::assertAttributeSame($expected, $attributeName, $actualObject, $message);
-            return;
-        }
-
-        // PHPUnit 9.0+.
-        try {
-            $actual = $this->getObjectAttributeValue($actualObject, $attributeName);
-        } catch (Exception $e) {
-            $this->fail($e->getMessage());
-        }
-
-        $this->assertSame($expected, $actual, $message);
-    }
-
-    /**
-     * Retrieve the value of an object's attribute.
-     * This also works for attributes that are declared protected or private.
-     *
-     * @param object|string $object        The object or class on which to check the property value.
-     * @param string        $attributeName The name of the property to check.
-     *
-     * @return mixed Property value.
-     *
-     * @throws \Exception
-     */
-    public static function getObjectAttributeValue($object, $attributeName)
-    {
-        $reflector = new ReflectionObject($object);
-
-        do {
-            try {
-                $attribute = $reflector->getProperty($attributeName);
-
-                if (!$attribute || $attribute->isPublic()) {
-                    return $object->$attributeName;
-                }
-
-                $attribute->setAccessible(true);
-                $value = $attribute->getValue($object);
-                $attribute->setAccessible(false);
-
-                return $value;
-            } catch (ReflectionException $e) {
-            }
-        } while ($reflector = $reflector->getParentClass());
-
-        throw new Exception(
-            \sprintf(
-                'Attribute "%s" not found in object.',
-                $attributeName
-            )
-        );
     }
 }
